requirements.txt
asgiref==3.7.2
Django==4.2.5
djangorestframework==3.14.0
joblib==1.3.2
numpy==1.26.0
opencv-python==4.8.0.76
pandas==2.1.1
Pillow==10.0.1
python-dateutil==2.8.2
python-magic==0.4.27
pytz==2023.3.post1
scikit-learn==1.3.1
scipy==1.11.2
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
